Moved it.SkipIf to skip.If and deprecated it.SkipIf; added slow condition
